+++ src/main/java/com/localstrategy/util/helper/CandleConstructor.java
@@ -1,29 +1,28 @@
 package com.localstrategy.util.helper;
 
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.Instant;
-import java.util.ArrayList;
-
 import com.localstrategy.util.types.Candle;
 import com.localstrategy.util.types.SingleTransaction;
 
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.ArrayList;
+
 public class CandleConstructor {
-    private double currentPrice;
-    private boolean candleToBeClosed;
+    private boolean closeOnNext;
     private double totalVolume;
     private double candleOpen;
     private double candleHigh;
     private double candleLow;
-    private double currentVolume;
-    private double volumePerCandle;
+    private final double volumePerCandle;
     private int transactionCount;
     private long lastTransactionId;
-    private ArrayList<Candle> candles = new ArrayList<>();
+    private final ArrayList<Candle> candles = new ArrayList<>();
     private int previousDayOfMonth = -1;
-    private int currentDayOfMonth;
-    private Candle candle;
     private int candleIndex = 0;
+
+    private double closeOnNextOpen;
+    private int closeOnNextTransactionCount;
 
     private static final int MAXIMUM_CANDLES_SIZE = 2000;
 
@@ -34,10 +33,10 @@ public class CandleConstructor {
     public Candle processTradeEvent(SingleTransaction transactionEvent){
 
         //TODO: We reset the total volume but don't adjust the prices? Look into it
-        currentDayOfMonth = LocalDateTime.ofInstant(
-                                Instant.ofEpochMilli(transactionEvent.getTimestamp()),
-                                ZoneId.systemDefault()
-                            ).getDayOfMonth();
+        int currentDayOfMonth = LocalDateTime.ofInstant(
+                Instant.ofEpochMilli(transactionEvent.timestamp()),
+                ZoneId.systemDefault()
+        ).getDayOfMonth();
 
         if(currentDayOfMonth != previousDayOfMonth){
             totalVolume = 0;
@@ -49,74 +48,44 @@ public class CandleConstructor {
         //TODO: What?
         Candle returnValue = null;
 
-        currentPrice = transactionEvent.getPrice();
+        double currentPrice = transactionEvent.price();
 
         //First transaction, candle volume larger than maximum or new day
+        Candle candle;
         if(totalVolume == 0){
 
             //If one transaction had volume larger than the maximum candle volume we must read the next transaction's price to set it as close
-            if(candleToBeClosed){
-                candle.setHigh(Math.max(candleHigh, currentPrice));
-                candle.setLow(Math.min(candleLow, currentPrice));
-                candle.setClose(currentPrice);
+            if(closeOnNext){
 
-                candle.setTimestamp(transactionEvent.getTimestamp()); //This is the time it closed TODO: this correct?
+                returnValue = makeAndGetCandle(transactionEvent, currentPrice, closeOnNextOpen, closeOnNextTransactionCount);
 
-                calculateDistance(candle);
-
-                if(candles.size() >= MAXIMUM_CANDLES_SIZE){
-                    candles.remove(0);
-                }
-
-                candles.add(candle);
-
-                returnValue = candle;
-
-                candleToBeClosed = false;
+                closeOnNext = false;
             }
-            candleOpen = currentPrice; //Perhaps set open price to this price only when starting from from 0th candle. Otherwise the open of the candle is not the close of the previous candle and that might be a problem
+            candleOpen = currentPrice; //Perhaps set open price to this price only when starting from 0th candle, otherwise the open of the candle is not the close of the previous candle and that might be a problem
             candleHigh = currentPrice;
             candleLow = currentPrice;
         }
 
-        currentVolume = transactionEvent.getAmount();
+        double currentVolume = transactionEvent.amount();
 
         if(totalVolume + currentVolume < volumePerCandle){
             transactionCount++;
             totalVolume += currentVolume;
             candleHigh = Math.max(currentPrice, candleHigh);
             candleLow = Math.min(currentPrice, candleLow);
-            //lastTransactionId = transactionEvent.getTransactionId();
+
+            //lastTransactionId = transactionEvent.getTransactionId(); //FIXME: Uncomment lastTransactionId during live testing
         } else {
             totalVolume = currentVolume > volumePerCandle ? 0 : currentVolume; //FIXME: This transfers volume over days as well
-            candleToBeClosed = currentVolume > volumePerCandle ? true : false;
+            closeOnNext = currentVolume > volumePerCandle;
 
-            candle = new Candle();
-            candle.setOpen(candleOpen);
-            candle.setHigh(Math.max(candleHigh, currentPrice));
-            candle.setLow(Math.min(candleLow, currentPrice));
-            candle.setClose(currentPrice);
+            if(!closeOnNext){ //If normal candle formed not a single transaction candle
 
-            candle.setVolume(transactionCount);
-            candle.setTimestamp(transactionEvent.getTimestamp());
-            candle.setIndex(candleIndex);
-            candleIndex++;
-            candle.setLastTransactionId(lastTransactionId); //OK it's the last transaction ID from the transaction IN the candle
+                returnValue = makeAndGetCandle(transactionEvent, currentPrice, candleOpen, transactionCount);
 
-            if(candleToBeClosed == false){ //If normal candle formed not a single transaction candle
-                calculateDistance(candle);
-
-                if(candles.size() >= MAXIMUM_CANDLES_SIZE){
-                    candles.remove(0);
-                }
-
-                candles.add(candle);
-
-                if(returnValue != null){
-                    System.out.println("obviously an error.");
-                }
-
-                returnValue = candle;
+            } else {
+                closeOnNextOpen = candleOpen;
+                closeOnNextTransactionCount = transactionCount;
             }
 
             candleOpen = currentPrice; //Correct - when volume is not higher than maximum volume - this transaction's price is current candle's group close and next candle's open.
@@ -129,7 +98,33 @@ public class CandleConstructor {
         return returnValue;
     }
 
-    public void calculateDistance(Candle candle){
+    private Candle makeAndGetCandle(SingleTransaction transactionEvent, double currentPrice, double candleOpen, int candleVolume) {
+
+        Candle candle;
+        candleHigh = Math.max(candleHigh, currentPrice);
+        candleLow = Math.min(candleLow, currentPrice);
+
+        candle = new Candle(
+                candleOpen,
+                candleHigh,
+                candleLow,
+                currentPrice,
+                candleVolume,
+                calculateDistance(candleHigh, candleLow),
+                candleIndex++,
+                transactionEvent.timestamp(),
+                lastTransactionId
+        );
+
+        if(candles.size() >= MAXIMUM_CANDLES_SIZE){
+            candles.remove(0);
+        }
+
+        candles.add(candle);
+        return candle;
+    }
+
+    public int calculateDistance(double candleHigh, double candleLow){
         int highDistance = 0;
         int lowDistance = 0;
 
@@ -140,10 +135,10 @@ public class CandleConstructor {
         for(; i < candles.size(); i++){
             Candle leftCandle = candles.get(candles.size()-1 - i);
 
-            if (highDistance == 0 && candle.getHigh() < leftCandle.getHigh()) {
+            if (highDistance == 0 && candleHigh < leftCandle.high()) {
                 highDistance = i;
             }
-            if (lowDistance == 0 && candle.getLow() > leftCandle.getLow()) {
+            if (lowDistance == 0 && candleLow > leftCandle.low()) {
                 lowDistance = -i;
             }
             if (highDistance != 0 && lowDistance != 0) {
@@ -161,14 +156,10 @@ public class CandleConstructor {
         }
 
         if(highDistance > -lowDistance){
-            candle.setTick(highDistance);
-        } else {
-            candle.setTick(lowDistance);
+            return highDistance;
         }
-    }
 
-    public double getCurrentPrice(){
-        return this.currentPrice;
+        return lowDistance;
     }
 
     public Candle getLastCandle(){
@@ -181,9 +172,5 @@ public class CandleConstructor {
 
     public ArrayList<Candle> getCandles(){
         return candles;
-    }
-
-    public void setCandles(ArrayList<Candle> candles){
-        this.candles = candles;
     }
 }