+++ src/main/java/com/localstrategy/util/types/UserDataStream.java
@@ -1,0 +1,17 @@
+package com.localstrategy.util.types;
+
+import com.localstrategy.Order;
+import com.localstrategy.UserAssets;
+
+import java.util.ArrayList;
+
+public record UserDataStream(UserAssets userAssets, ArrayList<Order> updatedOrders) {
+
+    public UserDataStream(
+            UserAssets userAssets,
+            ArrayList<Order> updatedOrders) {
+
+        this.userAssets = new UserAssets(userAssets);
+        this.updatedOrders = Order.deepCopyOrderList(updatedOrders);
+    }
+}